refactor client itself with standard adapter
