--- packages/client/src/rpc/index.ts
@@ -1,2 +1,0 @@
-export * from './json-serializer'
-export * from './serializer'
